fix code quality error (#581)
* fix code quality error

* try to fix compiler warning

+++ examples/tm_yolov4.cpp
@@ -162,7 +162,7 @@ std::vector<detection*> forward_darknet_layer_cpu(const float* input, layer l, i
     {
         int index = entry_index(l, i, 4, 0);
         logistic_cpu(l.output + index, l.w * l.h);
-        for (size_t loc = 0; loc < l.w * l.h; loc++)
+        for (size_t loc = 0; loc < (size_t)l.w * l.h; loc++)
         {
             if (l.output[index + loc] > s_thresh)
             {
